chatbot/requirements.txt
flask==2.0.1
openai==1.0.0
python-dotenv==0.19.1
requests==2.31.0
SQLAlchemy==1.4.23
